+++ pokemon_gen3/database/_connection.py
@@ -1,0 +1,13 @@
+from functools import lru_cache
+import sqlite3
+import pkg_resources
+from ..utils import atexit_context
+
+
+@lru_cache(None)
+def get_connection() -> sqlite3.Connection:
+    return atexit_context(
+        sqlite3.connect(
+            pkg_resources.resource_filename(__name__, "pokemon.db")
+        )
+    )