+++ pokemon_gen3/database/nature.py
@@ -1,12 +1,11 @@
 from typing import Iterator, cast
 
-from . import __connection
-
+from ._connection import get_connection
 from .._types import NatureID
 
 
 def ids() -> Iterator[NatureID]:
-    cursor = __connection.cursor()
+    cursor = get_connection().cursor()
     cursor.execute(
         """
 SELECT id FROM natures ORDER BY id
@@ -19,7 +18,7 @@ SELECT id FROM natures ORDER BY id
 def id_by_name_jp(
     name: str,
 ) -> NatureID:
-    cursor = __connection.cursor()
+    cursor = get_connection().cursor()
     cursor.execute(
         """
 SELECT id FROM natures WHERE name_jp=:name
@@ -33,7 +32,7 @@ SELECT id FROM natures WHERE name_jp=:name
 def name_jp_by_id(
     id_: NatureID,
 ) -> str:
-    cursor = __connection.cursor()
+    cursor = get_connection().cursor()
     cursor.execute(
         """
 SELECT name_jp FROM natures WHERE id=:id