+++ pokemon_gen3/database/stat.py
@@ -1,14 +1,13 @@
 import functools
-
 from typing import cast
 
-from . import __connection
+from ._connection import get_connection
 from .._types import StatID
 
 
 @functools.lru_cache()
 def name_en_by_id(id_: StatID) -> str:
-    cursor = __connection.cursor()
+    cursor = get_connection().cursor()
     cursor.execute(
         """
 SELECT name_en FROM stats WHERE id=:id
@@ -23,7 +22,7 @@ SELECT name_en FROM stats WHERE id=:id
 
 @functools.lru_cache()
 def name_jp_by_id(id_: StatID) -> str:
-    cursor = __connection.cursor()
+    cursor = get_connection().cursor()
     cursor.execute(
         """
 SELECT name_jp FROM stats WHERE id=:id