+++ tests/test_pokemon.py
@@ -8,9 +8,9 @@ def mew() -> Iterator[Pokemon]:
     yield Pokemon.from_name_jp("ミュウ")
 
 
-def test_mew_bases(mew: Pokemon) -> None:
-    assert len(mew.bases) == len(Stat)
-    assert all(base == 100 for base in mew.bases)
+def test_mew_base(mew: Pokemon) -> None:
+    assert len(mew.base) == len(Stat)
+    assert all(base == 100 for base in mew.base)
 
 
 def test_mew_name(mew: Pokemon) -> None: